Move validSyntax logic inside useStanc
